ref: move store to its own fowder
